Modify component structure, move CSS rules
