+++ src/Views/Selection/Selection.css
@@ -6,17 +6,3 @@
   align-items: center;
   padding: 3%;
 }
-
-.listFilters {
-  margin: 2%;
-  width: 90%;
-  border: 3px solid green;
-  text-align: center;
-  display: flex;
-  justify-content: space-evenly;
-  align-items: center;
-}
-
-.listFilters label {
-  font-size: larger;
-}
